GUI/requirements.txt
cobs==1.1.3
easygui==0.98.1
pyserial==3.2.1
